+++ algorithms/active/lstar/src/main/java/de/learnlib/algorithms/lstar/mealy/ClassicLStarMealy.java
@@ -25,14 +25,12 @@ import de.learnlib.algorithms.lstar.closing.ClosingStrategy;
 import de.learnlib.api.oracle.MembershipOracle;
 import de.learnlib.datastructure.observationtable.ObservationTable;
 import de.learnlib.datastructure.observationtable.Row;
-import de.learnlib.util.mealy.MealyUtil;
 import net.automatalib.automata.base.compact.CompactTransition;
 import net.automatalib.automata.concepts.SuffixOutput;
 import net.automatalib.automata.transducers.MealyMachine;
 import net.automatalib.automata.transducers.impl.compact.CompactMealy;
 import net.automatalib.words.Alphabet;
 import net.automatalib.words.Word;
-import org.checkerframework.checker.nullness.qual.Nullable;
 
 /**
  * An implementation of the L*Mealy algorithm for inferring Mealy machines, as described by Oliver Niese in his Ph.D.
@@ -46,7 +44,9 @@ import org.checkerframework.checker.nullness.qual.Nullable;
  * @author Malte Isberner
  */
 public class ClassicLStarMealy<I, O>
-        extends AbstractExtensibleAutomatonLStar<MealyMachine<?, I, ?, O>, I, @Nullable O, Integer, CompactTransition<O>, Void, O, CompactMealy<I, O>> {
+        extends AbstractExtensibleAutomatonLStar<MealyMachine<?, I, ?, O>, I, O, Integer, CompactTransition<O>, Void, O, CompactMealy<I, O>> {
+
+    private final O emptyOutput;
 
     /**
      * Constructor.
@@ -57,7 +57,7 @@ public class ClassicLStarMealy<I, O>
      *         the (Mealy) oracle
      */
     public ClassicLStarMealy(Alphabet<I> alphabet,
-                             MembershipOracle<I, @Nullable O> oracle,
+                             MembershipOracle<I, O> oracle,
                              ObservationTableCEXHandler<? super I, ? super O> cexHandler,
                              ClosingStrategy<? super I, ? super O> closingStrategy) {
         this(alphabet,
@@ -70,7 +70,7 @@ public class ClassicLStarMealy<I, O>
 
     @GenerateBuilder(defaults = AbstractExtensibleAutomatonLStar.BuilderDefaults.class)
     public ClassicLStarMealy(Alphabet<I> alphabet,
-                             MembershipOracle<I, @Nullable O> oracle,
+                             MembershipOracle<I, O> oracle,
                              List<Word<I>> initialPrefixes,
                              List<Word<I>> initialSuffixes,
                              ObservationTableCEXHandler<? super I, ? super O> cexHandler,
@@ -82,54 +82,29 @@ public class ClassicLStarMealy<I, O>
               LStarMealyUtil.ensureSuffixCompliancy(initialSuffixes, alphabet, true),
               cexHandler,
               closingStrategy);
-    }
-
-    public static <I, O> ClassicLStarMealy<I, O> createForSymbolOracle(Alphabet<I> alphabet,
-                                                                       MembershipOracle<I, @Nullable O> oracle,
-                                                                       ObservationTableCEXHandler<I, O> cexHandler,
-                                                                       ClosingStrategy<? super I, ? super O> closingStrategy) {
-        return new ClassicLStarMealy<>(alphabet, oracle, cexHandler, closingStrategy);
-    }
-
-    public static <I, O> ClassicLStarMealy<I, O> createForWordOracle(Alphabet<I> alphabet,
-                                                                     MembershipOracle<I, Word<O>> oracle,
-                                                                     ObservationTableCEXHandler<? super I, ? super O> cexHandler,
-                                                                     ClosingStrategy<? super I, ? super O> closingStrategy) {
-        return new ClassicLStarMealy<>(alphabet, MealyUtil.wrapWordOracle(oracle), cexHandler, closingStrategy);
+        this.emptyOutput = oracle.answerQuery(Word.epsilon());
     }
 
     @Override
-    protected MealyMachine<?, I, ?, O> exposeInternalHypothesis() {
+    public MealyMachine<?, I, ?, O> getHypothesisModel() {
         return internalHyp;
     }
 
     @Override
-    protected Void stateProperty(ObservationTable<I, @Nullable O> table, Row<I> stateRow) {
+    protected Void stateProperty(ObservationTable<I, O> table, Row<I> stateRow) {
         return null;
     }
 
     @Override
-    protected O transitionProperty(ObservationTable<I, @Nullable O> table, Row<I> stateRow, int inputIdx) {
+    protected O transitionProperty(ObservationTable<I, O> table, Row<I> stateRow, int inputIdx) {
         return table.cellContents(stateRow, inputIdx);
     }
 
     @Override
-    protected SuffixOutput<I, @Nullable O> hypothesisOutput() {
-        return new SuffixOutput<I, @Nullable O>() {
-
-            @Override
-            public @Nullable O computeOutput(Iterable<? extends I> input) {
-                return computeSuffixOutput(Collections.emptyList(), input);
-            }
-
-            @Override
-            public @Nullable O computeSuffixOutput(Iterable<? extends I> prefix, Iterable<? extends I> suffix) {
-                Word<O> wordOut = internalHyp.computeSuffixOutput(prefix, suffix);
-                if (wordOut.isEmpty()) {
-                    return null;
-                }
-                return wordOut.lastSymbol();
-            }
+    protected SuffixOutput<I, O> hypothesisOutput() {
+        return (prefix, suffix) -> {
+            final Word<O> wordOut = internalHyp.computeSuffixOutput(prefix, suffix);
+            return wordOut.isEmpty() ? emptyOutput : wordOut.lastSymbol();
         };
     }
 }