+++ algorithms/active/lstar/src/test/java/de/learnlib/algorithms/lstar/LearningTest.java
@@ -21,7 +21,6 @@ import de.learnlib.algorithms.lstar.closing.ClosingStrategies;
 import de.learnlib.algorithms.lstar.closing.ClosingStrategy;
 import de.learnlib.api.algorithm.LearningAlgorithm;
 import de.learnlib.api.oracle.EquivalenceOracle;
-import de.learnlib.api.oracle.MembershipOracle;
 import de.learnlib.api.query.DefaultQuery;
 import net.automatalib.automata.UniversalDeterministicAutomaton;
 import net.automatalib.words.Alphabet;
@@ -57,7 +56,6 @@ public class LearningTest {
             UniversalDeterministicAutomaton<?, I, ?, ?, ?> target,
             Alphabet<I> alphabet,
             LearningAlgorithm<M, I, D> learner,
-            MembershipOracle<I, D> oracle,
             EquivalenceOracle<? super M, I, D> eqOracle) {
         int maxRounds = target.size();