+++ algorithms/active/oml/src/main/java/de/learnlib/algorithms/oml/ttt/pt/PTNodeImpl.java
@@ -70,7 +70,7 @@ public class PTNodeImpl<I, D> implements PTNode<I, D> {
     }
 
     @Override
-    public PTNode<I, D> succ(I a) {
+    public @Nullable PTNode<I, D> succ(I a) {
         return children.get(a);
     }